@@ -9,6 +9,7 @@
   "author": "calocan <andy@likuski.org>",
   "license": "MIT",
   "main": "lib/index.js",
+  "type": "module",
   "module": "src/index.js",
   "files": [
     "lib",
@@ -19,7 +20,7 @@
     "clear-jest": "jest --clearCache",
     "clean": "rimraf lib dist es coverage",
     "lint": "eslint src",
-    "test": "cross-env BABEL_ENV=commonjs jest",
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
     "test:watch": "yarn test -- --watch",
     "test:cov": "yarn test -- --coverage",
     "build": "rollup -c",
@@ -54,7 +55,7 @@
     "@turf/turf": "^5.1.6",
     "@turf/union": "^6.0.3",
     "d3-scale": "^3.2.1",
-    "d3-scale-chromatic": "^1.5.0",
+    "d3-scale-chromatic": "^2.0.0",
     "folktale": "^2.3.2",
     "follow-redirects": "^1.10.0",
     "fs": "^0.0.1-security",
@@ -77,29 +78,29 @@
     "@babel/preset-env": "^7.8.4",
     "@babel/register": "^7.10.3",
     "@babel/runtime": "^7.8.4",
-    "@types/jest": "^25.1.3",
+    "@types/jest": "^26.0.15",
     "add": "^2.0.6",
     "babel-eslint": "^10.0.3",
-    "babel-jest": "^25.1.0",
+    "babel-jest": "^26.6.3",
     "babel-plugin-transform-es2015-modules-commonjs-simple": "^6.7.4",
     "babelrc-rollup": "^3.0.0",
     "cross-env": "^7.0.0",
     "dotenv": "8.2.0",
-    "eslint": "6.8.0",
-    "eslint-loader": "3.0.3",
-    "eslint-plugin-import": "2.20.1",
+    "eslint": "7.13.0",
+    "eslint-loader": "4.0.2",
+    "eslint-plugin-import": "2.22.1",
     "eslint-plugin-react": "^7.18.3",
-    "file-loader": "5.1.0",
-    "fs-extra": "8.1.0",
+    "file-loader": "6.2.0",
+    "fs-extra": "9.0.1",
     "install-peers-cli": "^2.2.0",
-    "jest": "^25.1.0",
+    "jest": "^26.6.3",
     "jsdoc-babel": "^0.5.0",
     "longjohn": "^0.2.12",
-    "meow": "^7.0.1",
+    "meow": "^8.0.0",
     "moment": "^2.24.0",
-    "promise": "8.0.3",
+    "promise": "8.1.0",
     "rimraf": "^3.0.2",
-    "rollup": "^1.31.1",
+    "rollup": "^2.33.1",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "10.1.0",
     "rollup-plugin-json": "^4.0.0",
@@ -107,7 +108,7 @@
     "rollup-plugin-node-globals": "1.4.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-terser": "^5.2.0",
+    "rollup-plugin-terser": "^7.0.2",
     "rollup-watch": "4.3.1"
   },
   "peerDependencies": {
@@ -133,12 +134,9 @@
       "<rootDir>/src/**/__tests__/**/*.{js,jsx,mjs}",
       "<rootDir>/src/**/?(*.)(spec|test).{js,jsx,mjs}"
     ],
-    "testEnvironment": "node",
+    "testEnvironment": "jest-environment-node",
     "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.(js|jsx|mjs)$": "<rootDir>/node_modules/babel-jest",
-      "^(?!.*\\.(js|jsx|mjs|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
+    "transform": {},
     "transformIgnorePatterns": [
       "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$"
     ],
