@@ -9,8 +9,7 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-// TODO jest can't read this file as babel :<
-const {rescapeDefaultTransports} = require('rescape-log');
+import {rescapeDefaultTransports} from 'rescape-log';
 
 // Set the loggers to debug level
 rescapeDefaultTransports.fileCombined.level = 'debug';
@@ -18,7 +17,7 @@ rescapeDefaultTransports.console.level = 'debug';
 
 // Enzyme setup
 // Set this to false to skip integration tests
-process.env.ENABLE_INTEGRATION_TESTS = process.env.ENABLE_INTEGRATION_TESTS  || 'false';
+process.env.ENABLE_INTEGRATION_TESTS = process.env.ENABLE_INTEGRATION_TESTS || 'false';
 if (process.env.ENABLE_INTEGRATION_TESTS !== 'false') {
   jest.unmock('query-overpass');
 }
@@ -26,8 +25,5 @@ if (process.env.ENABLE_INTEGRATION_TESTS !== 'false') {
 global.navigator = {
   userAgent: 'node.js'
 };
-if (process.env.NODE_ENV !== 'production') {
-  require('longjohn');
-}
 
 Error.stackTraceLimit = Infinity;
