@@ -1,7 +1,7 @@
 import babel from 'rollup-plugin-babel';
 import commonjs from 'rollup-plugin-commonjs';
 import pkg from './package.json';
-import * as R from 'ramda';
+import R from 'ramda';
 
 const config = {
   input: [
