@@ -28,9 +28,10 @@ import {
   resultToTaskWithResult, strPathOr
 } from 'rescape-ramda';
 import {loggers} from 'rescape-log';
-import * as R from 'ramda';
-import {of} from 'folktale/concurrency/task';
-import * as Result from 'folktale/result';
+import R from 'ramda';
+import T from 'folktale/concurrency/task';
+const {of} = T;
+import Result from 'folktale/result';
 
 const log = loggers.get('rescapeDefault');
 
