@@ -1,5 +1,5 @@
 import {defaultRunConfig} from 'rescape-ramda';
-import * as R from 'ramda';
+import R from 'ramda';
 import {queryLocationForOsmBlockOrAllResultsTask} from './overpassSingleOrAllBlocks';
 import {blocksToGeojson, blockToGeojson, locationsToGeojson} from './overpassBlockHelpers';
 
